Update habitat configuration and documentation for random state handling
- Set the default `torchDtype` to `float32` across various habitat configuration files to ensure consistency and improve compatibility with PyRadiomics.
- Revised comments and documentation to clarify the behavior of `random_state` parameters in different clustering modes, emphasizing inheritance and fallback mechanisms.
- Enhanced the `HabitatAnalysisConfig` class to include methods for resolving effective random states for individual and group clustering, improving clarity in random seed management.
- Updated user guide and configuration documentation to reflect these changes, ensuring users have clear guidance on the new defaults and their implications for clustering behavior.

+++ habit/utils/torch_radiomics_utils.py
@@ -22,7 +22,7 @@ TorchGpuSetting = Union[int, str, Sequence[Union[int, str]], None]
 
 DEFAULT_USE_TORCH_RADIOMICS = "auto"
 DEFAULT_TORCH_DEVICE = "auto"
-DEFAULT_TORCH_DTYPE = "float64"
+DEFAULT_TORCH_DTYPE = "float32"
 
 # Shown once per process when GPU TorchRadiomics is unavailable but would be used in auto mode.
 _TORCH_GPU_INSTALL_HINT_LOGGED = False
